# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Backend ::Deliah Restó:: App",
      "version": "1.0.0",
      "title": "Deliah Restó REST API",
      "contact": {
        "email": "carla.a.capocci@gmail.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "localhost:33333",
    "basePath": "/",
...
    "schemes": [
      "http"
    ],
    "paths": {
...
      "/usuarios": {
        "get": {
          "tags": [
            "Usuarios"
          ],
          "summary": "traer",
          "description": "trae todos los usuarios",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "Authorization",
              "description": "JWT Token",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "listado de usuarios",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Usuario"
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Usuarios"
          ],
          "summary": "crear",
          "description": "crea un usuario nuevo",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "datos",
              "description": "datos de usuario",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "usuario": {
                    "type": "string"
                  },
                  "nombre_y_apellido": {
                    "type": "string"
                  },
                  "telefono": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "direccion": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "ok",
              "schema": {
                "$ref": "#/definitions/ResponseOK"
              }
            }
          }
        }
      },
      "/usuarios/{id}": {
        "delete": {
          "tags": [
            "Usuarios"
          ],
          "summary": "eliminar",
          "description": "elimina un usuario",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "type": "string",
              "description": "Id del usuario"
            },
            {
              "in": "header",
              "name": "Authorization",
              "description": "JWT Token",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/ResponseOK"
              }
            }
          }
        },
        "put": {
          "tags": [
            "Usuarios"
          ],
          "summary": "modificar",
          "description": "modifica datos de un usuario",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "type": "string",
              "description": "Id del usuario"
            },
            {
              "in": "header",
              "name": "Authorization",
              "description": "JWT Token",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "datos",
              "description": "datos de usuario",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "usuario": {
                    "type": "string"
                  },
                  "nombre_y_apellido": {
                    "type": "string"
                  },
                  "telefono": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "direccion": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/ResponseOK"
              }
            }
          }
        },
        "patch": {
          "tags": [
            "Usuarios"
          ],
          "summary": "editar",
          "description": "solo el admin puede modificar los datos del usuario",
          "responses": {
            "200": {
              "description": "OK"
            }
          },
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "type": "string",
              "description": "Id del usuario"
            },
            {
              "in": "header",
              "name": "Authorization",
              "description": "JWT Token",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "datos",
              "description": "datos de usuario",
              "schema": {
                "type": "object",
                "properties": {
                  "usuario": {
                    "type": "string"
                  },
                  "nombre_y_apellido": {
                    "type": "string"
                  },
                  "telefono": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "direccion": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          ]
        }
      },
      "/usuarios/login": {
        "post": {
          "tags": [
            "Usuarios"
          ],
          "summary": "login",
          "description": "devuelve un token",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "credenciales",
              "description": "credenciales de usuario",
              "schema": {
                "type": "object",
                "properties": {
                  "usuario": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "login del usuario",
              "schema": {
                "$ref": "#/definitions/Login"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
      "Login": {
        "type": "object",
        "properties": {
          "loggedIn": {
            "type": "boolean"
          },
          "usuario": {
            "type": "object",
            "properties": {
              "usuario": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "rol": {
                "type": "string"
              }
            }
          },
          "token": {
            "type": "string"
          }
        }
      },
      "ResponseOK": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer"
          },
          "usuario": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "nombre_y_apellido": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "rol": {
            "type": "string"
          }
        }
      },
...
    }
  }